Fix core tests [29] --filter=[core] --verbose

+++ include/swoole_hash.h
@@ -19,7 +19,7 @@
 
 uint64_t swoole_hash_jenkins(const char *key, size_t keylen);
 uint64_t swoole_hash_php(const char *key, size_t len);
-uint64_t swoole_hash_austin(const char *key, unsigned int keylen);
-uint32_t swoole_crc32(const char *data, uint32_t size);
+uint64_t swoole_hash_austin(const char *key, size_t keylen);
+uint32_t swoole_crc32(const char *data, size_t size);
 
 #endif /* SW_HASH_H_ */
